requirements.txt
tensorflow>=2.10.0
numpy
pandas
matplotlib
scikit-learn
Pillow
streamlit
